Add function to retrieve more videos from infinite scroll

+++ playstv.py
@@ -1,12 +1,16 @@
+from asyncio.proactor_events import _ProactorBasePipeTransport
 from datetime import datetime
+from genericpath import isfile
 import json
 from operator import concat
 import os
 from pathlib import Path
 from time import sleep
-from typing import Annotated, Optional
+from typing import Annotated, Callable, Optional
+from urllib.parse import urlparse
 from bs4 import BeautifulSoup, NavigableString, Tag
 import requests
+from requests.models import PreparedRequest
 import typer
 from rich import print
 from rich.progress import track
@@ -77,6 +81,10 @@ class Video:
             file_name = concat(self.title, ".mp4")
             file_path = os.path.join(output_path, file_name)
 
+            if os.path.isfile(file_path):
+                structured_info("download", f"Download of {self.title} already exists!")
+                return
+
             video_stream = requests.get(self.video_url, stream=True)
 
             with open(file_path, "wb") as file:
@@ -126,19 +134,94 @@ class UserProfile:
         self.videos: [Video] = []
         self.original_url: str = str(f"https://plays.tv/u/{self.username}")
         self.archive_url: str = str()
+        self.__user_id__: str = str()
+        self.__module_url__: str = "https://plays.tv/ws/module"
         self.__last_video_fetched__: bool = False
         self.__last_video_id__: str = str()
         self.__archive_video_data__: [] = []
+        self.__current_page_number__: int = 0
         self.__wayback_user_agent__: str = str(
             "Mozilla/5.0 (Windows NT 5.1; rv:40.0) Gecko/20100101 Firefox/40.0"
         )
-        self.__archive_date__: datetime = datetime(year=2019, month=12, day=10, hour=17)
+        self.__archive_date__: datetime = datetime(year=2019, month=12, day=10)
 
     def __str__(self) -> str:
         return f"---------\nUsername: {self.username} \nPossible videos: {len(self.video_ids)} \nActual Videos {len(self.videos)} \n---------"
 
     def get_id_from_url(url: str) -> str:
         pass
+
+    def get_user_id(self) -> None:
+        try:
+            content = requests.get(self.archive_url).content
+            soup = BeautifulSoup(content, "html.parser")
+            html_element = soup.find("button", {"title": "Add Friend"})
+
+            if html_element is not None:
+                self.__user_id__ = html_element.attrs["data-obj-id"]
+                structured_info(
+                    "initialization", f"Retrieved user id ({self.__user_id__})"
+                )
+        except:
+            structured_error("initialization", "Failed to retrieve user id")
+
+    def check_video_availability(self) -> None:
+        for vid_index in track(
+            range(len(self.videos)), description="Checking video availability..."
+        ):
+            current_vid: Video = self.videos[vid_index]
+            current_vid.check_availability()
+
+    def get_more_videos(self) -> list[Video] | None:
+        if self.__last_video_fetched__:
+            self.check_video_availability()
+            return
+        video_list: list[Video] = []
+
+        params = {
+            "infinite_scroll": True,
+            "infinite_scroll_fire_only": True,
+            "custom_loading_module_state": "appending",
+            "page_num": self.__current_page_number__ + 1,
+            "target_user_id": self.__user_id__,
+            "last_id": self.__last_video_id__,
+            "section": "videos",
+            "format": "application/json",
+            "id": "UserVideosMod",
+        }
+
+        module_url = PreparedRequest()
+        module_url.prepare_url(self.__module_url__, params)
+
+        parsed_archive_url = urlparse(self.archive_url)
+
+        #
+
+        query_url = f"{parsed_archive_url.scheme}://{parsed_archive_url.netloc}/{'/'.join(parsed_archive_url.path.split('/')[1:3])}/{module_url.url}"
+
+        r = requests.get(query_url)
+
+        json_data = json.loads(r.text)
+
+        if json_data["body"] == "":
+            self.__last_video_fetched__ = True
+            return
+
+        soup_body = BeautifulSoup(json_data["body"], "html.parser")
+
+        # TODO: Parse videos from json body attribute
+
+        # foreach .video-item get href from info div
+        # after retrieving this, take video_id from href (second element in path)
+        # create embeds url for this video_id
+        # in case recursive function case is hit, check availability of all videos
+        # after which the recursive function ends
+        # ONLY ADD UNIQUE ID
+        self.videos += video_list
+
+        self.__current_page_number__ += 1
+        # TODO: Test recursive function
+        self.get_more_videos()
 
     def download_videos(self, output_path: str) -> None:
         for video_index in track(range(len(self.videos)), "Downloading videos..."):
@@ -157,7 +240,6 @@ class UserProfile:
             year=self.__archive_date__.year,
             month=self.__archive_date__.month,
             day=self.__archive_date__.day,
-            hour=self.__archive_date__.hour,
         )
 
         try:
@@ -178,7 +260,6 @@ class UserProfile:
             year=possible_url_timestamp.year,
             month=possible_url_timestamp.month,
             day=possible_url_timestamp.day,
-            hour=possible_url_timestamp.hour,
         )
 
         if self.__archive_date__ != possible_url_datetime:
@@ -189,7 +270,8 @@ class UserProfile:
         self.archive_url = possible_url.archive_url
         return True
 
-    def get_initial_video_ids(self) -> None:
+    def get_initial_videos(self) -> None:
+        structured_info("")
         response = requests.get(self.archive_url)
         soup = BeautifulSoup(response.content, "html.parser")
 
@@ -241,34 +323,3 @@ def structured_error(subject: str, message: str):
 
 def structured_warning(subject: str, message: str):
     structured_print(f"warning - {subject}", message, "bright_yellow")
-
-
-app = typer.Typer()
-
-
-@app.command()
-def main(output_path: Annotated[Optional[Path], typer.Option()]):
-    if not output_path.is_dir():
-        structured_error("initialization", "Specified path is not a directory")
-        return
-
-    user = typer.prompt("What's your Plays.tv username?")
-    me = get_profile(user)
-
-    if me is None:
-        structured_error("initialization", "Could not get user profile")
-        return
-
-    me.get_initial_video_ids()
-
-    for vid_index in track(
-        range(len(me.videos)), description="Checking video availability..."
-    ):
-        current_vid: Video = me.videos[vid_index]
-        current_vid.check_availability()
-
-    me.download_videos(output_path.name)
-
-
-if __name__ == "__main__":
-    app()
